pyproject.toml
[project]
name = "scrapping-file-word"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "python-docx>=1.2.0",
    "tabulate>=0.9.0",
]
